% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_connectivity-package.R
\docType{package}
\name{river_connectivity-package}
\alias{river_connectivity}
\alias{river_connectivity-package}
\title{river_connectivity: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jacquetclaire/river_connectivity}
  \item Report bugs at \url{https://github.com/jacquetclaire/river_connectivity/issues}
}

}
\author{
\strong{Maintainer}: Claire Jacquet \email{claire.jacquet@umontpellier.fr} (\href{https://orcid.org/0000-0002-7913-0689}{ORCID}) [copyright holder]

}
\keyword{internal}
